Extracted an Allocator interface for DirectByteBuffers
