+++ jme3-core/src/main/java/com/jme3/util/BufferAllocator.java
@@ -1,0 +1,14 @@
+package com.jme3.util;
+
+import java.lang.reflect.Method;
+import java.nio.Buffer;
+import java.nio.ByteBuffer;
+import java.nio.DoubleBuffer;
+
+public interface BufferAllocator {
+
+	void destroyDirectBuffer(Buffer toBeDestroyed);
+
+	ByteBuffer allocate(int size);
+
+}
